+++ src/main/java/com/petCart/service/IProductService.java
@@ -12,8 +12,8 @@ public interface IProductService {
     public List<Product> getAllProduct();
 	public void deleteProduct(Product product);
 	public Product getProduct();
-	public Product findById(Integer id);
-	public Product viewProduct(Integer id);
+	public Product findById(Long id);
+	public Product viewProduct(Long id);
 	public List<Product> search(SearchContext context,Integer lowerLimit,Integer upperLimit, String orderBy, String orderType);
